Linux build fix: Fix C++17 int64_t cast in mcp_example_server ping handler (#124)
Change long long cast to int64_t for timestamp parameter in metadata
to fix C++17 variant compatibility issue.

+++ examples/mcp/mcp_example_server.cc
@@ -656,7 +656,7 @@ void setupServer(McpServer& server, bool verbose) {
           auto pong =
               make<Metadata>()
                   .add("pong", true)
-                  .add("timestamp", static_cast<long long>(std::time(nullptr)))
+                  .add("timestamp", static_cast<int64_t>(std::time(nullptr)))
                   .build();
 
           return jsonrpc::Response::success(request.id,
